requirements.txt
# BERT Production Optimization - Complete Dependencies

# Core ML and Deep Learning
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Google Cloud Platform (Vertex AI GPU Service)
google-cloud-aiplatform>=1.111.0
google-cloud-storage>=2.19.0
google-cloud-compute>=1.14.0
google-cloud-billing>=1.12.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0

# Configuration and Data Handling
pyyaml>=6.0.2
dataclasses-json>=0.5.9

# System Monitoring and Utilities
psutil>=5.9.0
tqdm>=4.65.0

# CLI and User Interface
click>=8.1.0
rich>=13.4.0

# Logging and Monitoring
structlog>=23.1.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.23.0